Keep caller-supplied pages and line ranges from overflowing (#233)
A page is clamped only from below, so page times page size wrapped an
int negative: DuckDB refused the OFFSET as an error on grep, git_log and
/commits, and the multiline scan answered page 1 labelled as the page
asked for. Offsets are now computed as long, so the largest page is a
page past the end like any other.

Glob paging counted the offset in the asked page size while each page
held the clamped one, so pageSize=5000 skipped rows 2001 to 4999 and
echoed 5000 back. IndexReader.GlobAsync now takes the page and cuts it
with the clamped size, and the listing echoes the size it used.

read_file matched ranges with \d, which accepts any Unicode digit, and
int.Parse then threw on those and on numbers past int.MaxValue, failing
the whole call. The ranges match [0-9] and parse with TryParse; a number
too large is a refusal of that entry alone.

+++ CodeExplorer/Search/CommitLogTools.cs
@@ -76,7 +76,7 @@ internal sealed partial class HistoryTools
         // end of a real author's commits are opposite facts and the second sentence would fit both.
         if (answer.Author is { Addresses: 0 } miss) return NoSuchAuthor(miss, answer.Repository, answer.Path);
 
-        int skip = (answer.Page - 1) * answer.Limit;
+        long skip = (answer.Page - 1L) * answer.Limit;
         string where = Scope(answer.Repository, answer.Path);
         if (answer.Commits.Count == 0)
             return (skip > 0
